// src/app/business-logic/model/projects/metricVariantResult.ts
/**
 * projects
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.9
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface MetricVariantResult {
    /**
     * Metric name
     */
    metric?: string;
    /**
     * Metric name hash. Used instead of the metric name when categorizing                 last metrics events in task objects.
     */
    metric_hash?: string;
    /**
     * Variant name
     */
    variant?: string;
    /**
     * Variant name hash. Used instead of the variant name when categorizing                 last metrics events in task objects.
     */
    variant_hash?: string;
}